コンポーネントからdbを直接扱わないように
