--- middleware/auth.js
@@ -1,5 +1,0 @@
-export default function ({ route, redirect, store }) {
-  if (route.fullPath !== '/' && !store.getters.userSignedIn) {
-    return redirect('/')
-  }
-}
