+++ middleware/auth.ts
@@ -1,0 +1,7 @@
+import { Context } from '@nuxt/types'
+
+export default function ({ route, redirect, store }: Context) {
+  if (route.fullPath !== '/' && !store.getters['users/userSignedIn']) {
+    return redirect('/')
+  }
+}
