+++ models/index.ts
@@ -1,2 +1,2 @@
-export { default as Note } from './note'
+export { Note } from './note'
 export { default as User } from './user'
